+++ docs/CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.11.3 (Unreleased)
+-------------------
+- Enh: Automated code refactoring for HumHub 1.18 using Rector
+
+0.11.2 (February 24, 2026)
+--------------------------
+- Enh: Automated code refactoring for HumHub 1.18 using Rector
+- Enh #232: Add custom notification endpoint
+
 0.11.1 (November 19, 2025)
 --------------------------
 - Enh #219: Improved Module Test GitHub Actions